IV-1472 Add setup automatic upgrade script

+++ rlw/setup_hostname.ps1
@@ -10,5 +10,9 @@ if ($env:SERVER_HOSTNAME -and $env:SERVER_HOSTNAME -ne (& hostname)) {
   Rename-Computer $env:SERVER_HOSTNAME
 
   # Host Name does not take effect until after a computer restart
-  Restart-Computer -Force
+  if ($env:WINDOWS_UPDATES_REBOOT_SETTING -eq "Allow Reboot") {
+    Restart-Computer -Force
+  } else {
+    return "Reboot is required, but not allowed by WINDOWS_UPDATES_REBOOT_SETTING input. Please do the reboot manually."
+  }
 }
